Fix pull requests commit messages checks
By default, all commit messages in PRs will be checked to ensure that
the lines length don't exceed 72 characters.

Also, ensure that the title or description of the PR includes a link to
the issue that has been fixed/resolved. Allowed patterns are:

* Fixes AlmaLinux/build-system#1234
* Resolves: https://github.com/AlmaLinux/build-system/issues/1234

Note that it checks the PR title/description, which by default is
taken from the first commit added into the PR. This way we ensure that:

* In case of a single commit PR, the commit message includes a
  resolves/fixes reference.
* In case of multiple commits PR, the first message might not include
  such reference, and it must be added manually when merging the PR into
  the main branch. Note that, the title or description of the PR must
  include such reference, which can be updated manually anytime.

Fixes: https://github.com/AlmaLinux/build-system/issues/146